+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Typescript migration
 - Upgrade React to v16.9+
 - react-testing-library for test
+- use react-router 5+ hooks
 
 # v7.1.0